added upgrade multipliers for gauge upgrades

+++ core/src/com/udacity/gamedev/gigagal/entity/Humanoid.java
@@ -9,7 +9,7 @@ public interface Humanoid extends MultidirectionalX, MultidirectionalY {
     Vector2 getVelocity();
     Rectangle getBounds();
     float getTurbo();
-    int getHealth();
+    float getHealth();
     boolean getJumpStatus();
     boolean getHoverStatus();
     boolean getClingStatus();
